Move ContourHelper to shared project. Move MeshGeometryHelper to shared project.
